Move hocs used on frontend to base
