+++ assets/js/base/hocs/with-rest-api-hydration.js
