+++ assets/js/blocks/attribute-filter/frontend.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { withRestApiHydration } from '@woocommerce/block-hocs';
+import withRestApiHydration from '@woocommerce/base-hocs/with-rest-api-hydration';
 
 /**
  * Internal dependencies
