+++ assets/js/blocks/cart-checkout/cart/frontend.js
@@ -1,10 +1,8 @@
 /**
  * External dependencies
  */
-import {
-	withRestApiHydration,
-	withStoreCartApiHydration,
-} from '@woocommerce/block-hocs';
+import withRestApiHydration from '@woocommerce/base-hocs/with-rest-api-hydration';
+import withStoreCartApiHydration from '@woocommerce/base-hocs/with-store-cart-api-hydration';
 import { __ } from '@wordpress/i18n';
 import { useStoreCart } from '@woocommerce/base-hooks';
 import { RawHTML } from '@wordpress/element';
