+++ assets/js/hocs/index.js
@@ -6,5 +6,3 @@ export { default as withProduct } from './with-product';
 export { default as withProductVariations } from './with-product-variations';
 export { default as withSearchedProducts } from './with-searched-products';
 export { default as withTransformSingleSelectToMultipleSelect } from './with-transform-single-select-to-multiple-select';
-export { default as withRestApiHydration } from './with-rest-api-hydration';
-export { default as withStoreCartApiHydration } from './with-store-cart-api-hydration';
